fix: Sat modifier has value of 0.2 but I see in View that it's upgraded for 1 (0.1*10)     In battle against knight on 3xbuff it goes 10->12->13->16

+++ Assets/Core/Game/GamePlay/Scripts/Presentation/Stat/StatItemData.cs
@@ -6,15 +6,15 @@ namespace Presentation.Stat
     {
         public int Id { get; }
         public StatType StatType { get; }
-        public int BaseValue { get; }
-        public int CurrentValue { get; }
+        public float BaseValue { get; }
+        public float CurrentValue { get; }
 
 
         public StatItemData(
             int id,
             StatType statType,
-            int baseValue,
-            int currentValue)
+            float baseValue,
+            float currentValue)
         {
             Id = id;
             StatType = statType;
